Add bindable query parameters using Encodable
Adds a new type `Insert<Table, Parameters>` which models a cached
prepared statement that can be re-executed multiple times with different
bound parameters. Call `bind()` with a set of parameters to receive a
`Query<Int64>` that can be executed via the `Database`, optionally
producing the `Int64` rowids that were inserted.

In order to facilitate a more declarative style of query definition,
adds a `Database.current` helper that looks for a database for the
current queue, so that static query definitions can lazily reference
the "current" database at the time they're first accessed.